+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/ImpTest.hs
@@ -839,6 +839,7 @@ instance
             sgGenDelegs = mempty
           , sgInitialFunds = mempty
           , sgStaking = mempty
+          , sgExtraConfig = SNothing
           }
     case validateGenesis gen of
       Right () -> pure gen